test(ui): проверка снятой галки перестала засчитывать её исчезновение (#888)
Тест утверждал «если ключ в ответе есть, он не true» — то есть отсутствие поля
в ответе считал успехом. А отсутствие хуже: клиент применяет values к форме и
ключи, которых в ответе нет, не трогает. Галка осталась бы взведённой на
экране, хотя пользователь её снял, — ровно то поведение, ради недопущения
которого тест и писался.

Теперь проверяется полный контракт: поле обязано БЫТЬ в ответе и быть ложью;
парная проверка — взведённая галка возвращается взведённой (иначе «всегда
ложь» прошло бы обе). Значение трактуется как у клиента, поиск ключа
регистронезависимый: Object.Set хранит имена в нижнем регистре.

Мутационные проверки: удаление ключа из ответа роняет оба теста с объяснением;
принудительная ложь роняет парный.

Файл переименован из zz_verify_checkbox_test.go — имя осталось от разовой
проверки, а тест давно постоянный.

Generated-with: Claude Code
